requirements.txt
certifi==2021.5.30
chardet==4.0.0
config-parser==0.0.1
configparser==5.0.2
idna==2.10
kafka-python==2.0.2
pynput==1.7.3
PyYAML==5.4.1
six==1.16.0
urllib3==1.26.6
